+++ limberframework/authentication/authentication_service_provider.py
@@ -11,9 +11,16 @@ from limberframework.support.services import ServiceProvider
 class AuthServiceProvider(ServiceProvider):
     """Registers database services to the service container."""
 
-    def register(self) -> None:
+    def register(self, app: Application) -> None:
+        """Registers the auth service with the application.
+
+        Arguments:
+        app limberframework.foundation.application.Application --
+        the service container.
+        """
+
         async def register_authenticator(app: Application):
             config_service = await app.make("config")
             return await make_authenticator(config_service["auth"])
 
-        self.app.bind("auth", register_authenticator, defer=True)
+        app.bind("auth", register_authenticator, defer=True)