+++ limberframework/config/config_service_provider.py
@@ -11,9 +11,13 @@ from limberframework.support.services import ServiceProvider
 class ConfigServiceProvider(ServiceProvider):
     """Registers configuration services to the service container."""
 
-    def register(self) -> None:
+    def register(self, app: Application) -> None:
         """Registers the Config class to the service container,
         which holds configuration settings for the application.
+
+        Arguments:
+        app limberframework.foundation.application.Application --
+        the service container.
         """
 
         async def register_config(app: Application) -> Config:
@@ -27,4 +31,4 @@ class ConfigServiceProvider(ServiceProvider):
             """
             return Config()
 
-        self.app.bind("config", register_config, True)
+        app.bind("config", register_config, True)