+++ limberframework/database/database_service_provider.py
@@ -13,9 +13,13 @@ from limberframework.support.services import ServiceProvider
 class DatabaseServiceProvider(ServiceProvider):
     """Registers database services to the service container."""
 
-    def register(self) -> None:
+    def register(self, app: Application) -> None:
         """Registers the database connection and session
         services to the service container.
+
+        Arguments:
+        app limberframework.foundation.application.Application --
+        the service container.
         """
 
         async def register_database_connection(app: Application) -> Connection:
@@ -30,7 +34,7 @@ class DatabaseServiceProvider(ServiceProvider):
             config_service = await app.make("config")
             return await make_connection(config_service["database"])
 
-        self.app.bind("db.connection", register_database_connection)
+        app.bind("db.connection", register_database_connection)
 
         async def register_database_session(app: Application) -> Session:
             """Closure for establishing a database session
@@ -45,4 +49,4 @@ class DatabaseServiceProvider(ServiceProvider):
             db_connection = await app.make("db.connection")
             return sessionmaker(bind=db_connection.engine)()
 
-        self.app.bind("db.session", register_database_session, defer=True)
+        app.bind("db.session", register_database_session, defer=True)